jar 파일 전달
