+++ server/src/main/java/com/example/neckisturtle/feature/dto/SignupDto.java
@@ -1,0 +1,2 @@
+package com.example.neckisturtle.feature.dto;public class SignupDto {
+}
