+++ server/src/main/java/com/example/neckisturtle/feature/persistance/MissionRecordRepo.java
@@ -1,0 +1,2 @@
+package com.example.neckisturtle.feature.persistance;public interface MissionRecordRepo {
+}
