+++ server/src/main/java/com/example/neckisturtle/feature/service/MissionService.java
@@ -1,0 +1,2 @@
+package com.example.neckisturtle.feature.service;public class MissionService {
+}
