--- server/src/main/resources/META-INF/MANIFEST.MF
@@ -1,3 +1,0 @@
-Manifest-Version: 1.0
-Main-Class: com.example.neckisturtle.NeckIsTurtleApplication
-
